created the best AI yet.
created infantAI, the smartest AI yet, through a hideous, disgusting
gelatinous blob of spaghetti code.  However, it works and functions as
intended.  I will attempt to clean it up later and make it non-hacky.

infantAI plays exactly like dumbAI except that it will not place a piece
where the opponent can play immediately on top of the piece played to
connect 4 in a row or prevent infantAI from making 4 in a row.